Separate NoopStream and BufferStream
