+++ lib/parse.js
@@ -4,6 +4,8 @@ var Stream = require('stream');
 var binary = require('binary');
 var Promise = require('bluebird');
 var PullStream = require('./PullStream');
+var NoopStream = require('./NoopStream');
+var BufferStream = require('./BufferStream');
 
 // Backwards compatibility for node 0.8
 if (!Stream.Writable)
@@ -12,16 +14,6 @@ if (!Stream.Writable)
 var endDirectorySignature = new Buffer(4);
 endDirectorySignature.writeUInt32LE(0x06054b50, 0);
 
-function NoopStream() {
-  if (!(this instanceof NoopStream)) {
-    return new NoopStream();
-  }
-  Stream.Transform.call(this);
-}
-util.inherits(NoopStream,Stream.Transform);
-
-NoopStream.prototype._transform = function(d,e,cb) { cb() ;};
-  
 function Parse(opts) {
   if (!(this instanceof Parse)) {
     return new Parse(opts);
@@ -95,20 +87,7 @@ Parse.prototype._readFile = function () {
       };
 
       entry.buffer = function() {
-        return new Promise(function(resolve,reject) {
-          var buffer = new Buffer(''),
-              bufferStream = Stream.Transform()
-                .on('finish',function() {
-                  resolve(buffer);
-                })
-                .on('error',reject);
-              
-          bufferStream._transform = function(d,e,cb) {
-            buffer = Buffer.concat([buffer,d]);
-            cb();
-          };
-          entry.pipe(bufferStream);
-        });
+        return BufferStream(entry);
       };
 
       entry.path = fileName;
